Add files via upload
This is a Data Analytics Project on  T20I WC 2022,the analysis shows the best 11 in the T20I format. Not only that it also shows their individual stats, a comparison of their stats is done of all the players playing the T20I format of the sport and according to stats the best 11 is selected. Also there are different slicers that show batters batting at different position having different stats and bowlers with the best stats in contribution of both with the bat and the ball have been selected in the playing 11.